--- src/main/java/com/techflow/techflow/dto/LoginDto.java
@@ -1,23 +1,0 @@
-package com.techflow.techflow.dto;
-
-public class LoginDto {
-    private String email;
-
-    private String password;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
