--- src/main/java/com/techflow/techflow/dto/LoginResponse.java
@@ -1,22 +1,0 @@
-package com.techflow.techflow.dto;
-
-public class LoginResponse {
-    private String token;
-    private long expiresIn;
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public long getExpiresIn() {
-        return expiresIn;
-    }
-
-    public void setExpiresIn(long expiresIn) {
-        this.expiresIn = expiresIn;
-    }
-}
