--- src/main/java/com/techflow/techflow/dto/RegisterDto.java
@@ -1,36 +1,0 @@
-package com.techflow.techflow.dto;
-
-import com.techflow.techflow.validator.StrongPassword;
-
-public class RegisterDto {
-    private String email;
-
-    @StrongPassword
-    private String password;
-
-    private String fullName;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getFullName() {
-        return fullName;
-    }
-
-    public void setFullName(String fullName) {
-        this.fullName = fullName;
-    }
-}
