+++ src/main/java/com/techflow/techflow/dto/request/LoginDto.java
@@ -1,0 +1,20 @@
+package com.techflow.techflow.dto.request;
+
+import com.techflow.techflow.validator.StrongPassword;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.Getter;
+import lombok.NonNull;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginDto {
+    @NotBlank(message = "L'email est obligatoire")
+    @Email(message = "L'email doit être valide")
+    private String email;
+    @StrongPassword
+    private String password;
+
+}
