+++ src/main/java/com/techflow/techflow/dto/response/LoginResponse.java
@@ -1,0 +1,15 @@
+package com.techflow.techflow.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginResponse {
+    private String token;
+    private long expiresIn;
+}
