+++ src/main/java/com/techflow/techflow/repository/UtilisateurRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UtilisateurRepository extends CrudRepository<Utilisateur, Integer> {
+public interface UtilisateurRepository extends CrudRepository<Utilisateur, String> {
     Optional<Utilisateur> findByEmail(String email);
 }
