+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.application.name=techflow
 spring.datasource.url=jdbc:postgresql://localhost:5432/techflow
-spring.datasource.username=root
-spring.datasource.password=root
+spring.datasource.username=postgres
+spring.datasource.password=
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 spring.datasource.initialize=true
